feat: ensure cart product quantity can be incrased and decrased

+++ src/action/decrease-cart-product-quantity/index.ts
@@ -8,13 +8,13 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
-import { RemoveProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
+import { DecreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: RemoveProductFromCartSchema,
+export const decreaseCartProductQuantity = async (
+  data: DecreaseCartProductQuantitySchema,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -34,5 +34,15 @@ export const removeProductFromCart = async (
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  } else {
+    await db
+      .update(cartItemTable)
+      .set({
+        quantity: cartItem.quantity - 1,
+      })
+      .where(eq(cartItemTable.id, cartItem.id));
+  }
 };
